+++ src/group/symmetric.rs
@@ -8,7 +8,7 @@ use crate::group::cycle::Cycle;
 pub fn sym(n_size: i32) -> SymmetricGroup<i32> {
     let ground: Vec<i32> = (1..=n_size).collect();
     SymmetricGroup {
-        n: n_size,
+        n: n_size as usize,
         ground: ground,
     }
 }
@@ -16,14 +16,14 @@ pub fn sym(n_size: i32) -> SymmetricGroup<i32> {
 #[derive(Default)]
 pub struct SymmetricGroup<V> {
     /// the base size of a permutation the order should be n!
-    n: i32,
+    n: usize,
     ground: Vec<V>,
 }
 impl<V> SymmetricGroup<V> 
 where
     V: Clone+Copy+Hash+Eq+PartialEq+Debug
 {
-    pub fn new(n_size: i32, g: Vec<V>) -> SymmetricGroup<V> {
+    pub fn new(n_size: usize, g: Vec<V>) -> SymmetricGroup<V> {
         SymmetricGroup {
             n: n_size,
             ground: g,
@@ -70,7 +70,7 @@ where
     /// The order of a symmetric group means the size.
     /// Computing the size is too hard, so we will just use the following formula instead.
     fn order(&self) -> i32 {
-        (1..=self.n).product()
+        (1..=self.n as i32).product()
     }
 
     /// In Symmetric Groups, we can generate every single n! possible permutations by only combining and adding adjacent transpositions, ie (a_i, a_{i + 1}) for all i.
@@ -93,7 +93,7 @@ mod tests {
     #[test]
     fn test_symmetric_3() { 
         let ground: Vec<i32> = vec![1, 2, 3];
-        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len() as i32, ground.clone());
+        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len(), ground.clone());
         let symmetric_set = group.get_set();
         
         debug_assert_eq!(symmetric_set.contains(&Cycle::from(vec![vec![1, 2, 3]], ground.clone())), true);
@@ -106,7 +106,7 @@ mod tests {
     #[test]
     fn test_symmetric_5() { 
         let ground: Vec<i32> = vec![1, 2, 3, 4, 5];
-        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len() as i32, ground.clone());
+        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len(), ground.clone());
         let symmetric_set = group.get_set();
         debug_assert_eq!(group.order(), symmetric_set.len() as i32);
         debug_assert_eq!(group.order(), 120);
@@ -115,7 +115,7 @@ mod tests {
     #[test]
     fn test_symmetric_8() { 
         let ground: Vec<i32> = vec![1, 2, 3, 4, 5, 6, 7, 8];
-        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len() as i32, ground.clone());
+        let group: SymmetricGroup<i32> = SymmetricGroup::new((&ground).len(), ground.clone());
         let symmetric_set = group.get_set();
         debug_assert_eq!(40320, symmetric_set.len() as i32);
     }
